Crate sui_indexer_alt_jsonrpc

Modules

api

args

config

data

Structs

NodeArgs

RpcArgs

RpcService

Functions

start_rpc

Set-up and run the RPC service, using the provided arguments (expected to be extracted from the command-line). The service will continue to run until the cancellation token is triggered, and will signal cancellation on the token when it is shutting down.