sui_indexer_alt_jsonrpc::api::write
Trait WriteApiClient

pub trait WriteApiClient: ClientT {
    // Provided methods
    fn execute_transaction_block<'life0, 'async_trait>(
        &'life0 self,
        tx_bytes: Base64,
        signatures: Vec<Base64>,
        options: Option<SuiTransactionBlockResponseOptions>,
        request_type: Option<ExecuteTransactionRequestType>,
    ) -> Pin<Box<dyn Future<Output = Result<SuiTransactionBlockResponse, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn dry_run_transaction_block<'life0, 'async_trait>(
        &'life0 self,
        tx_bytes: Base64,
    ) -> Pin<Box<dyn Future<Output = Result<DryRunTransactionBlockResponse, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
}

Client implementation for the WriteApi RPC API.

Provided Methods

fn execute_transaction_block<'life0, 'async_trait>( &'life0 self, tx_bytes: Base64, signatures: Vec<Base64>, options: Option<SuiTransactionBlockResponseOptions>, request_type: Option<ExecuteTransactionRequestType>, ) -> Pin<Box<dyn Future<Output = Result<SuiTransactionBlockResponse, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Execute the transaction with options to show different information in the response. The only supported request type is WaitForEffectsCert: waits for TransactionEffectsCert and then return to client. WaitForLocalExecution mode has been deprecated.

fn dry_run_transaction_block<'life0, 'async_trait>( &'life0 self, tx_bytes: Base64, ) -> Pin<Box<dyn Future<Output = Result<DryRunTransactionBlockResponse, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return transaction execution effects including the gas cost summary, while the effects are not committed to the chain.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.